refactor(report): split _suppress_dangling_correlation_notes into its own leaf module
reporter_markdown.py hit the AI-readiness 2000-line hard cap (2008
lines) after the previous two commits added --show-only correlation
cleanup shared across every report format. The helper was never
markdown-specific to begin with -- it's called by JSON, SARIF, HTML,
and JUnit too -- so it belongs in its own shared leaf module rather
than growing the markdown renderer further.

Moved to abicheck/report_correlation.py (imports only from .checker,
same leaf-module discipline reporter_markdown.py documents for
itself). reporter_markdown.py now imports and re-exports it under the
same name via the standard `as`-aliased explicit-reexport pattern
(mypy flagged the first attempt without it -- a plain `from .x import
y` isn't an explicit reexport reporter.py's own `as`-aliased
re-import can see). No functional change and no import-path change
for any of the four downstream formats.
